aggregator: (pre-)move availability controller
